New Code Outline
I created a general outline of the code and some of the function prototypes. I also initialized all the pins we will be using and started the main loop. The main loop detects which buttons are pressed, and depending on the button pressed, the corresponding functions for each parameter readings are called